third_party/blink/web_tests/accessibility/inline-text-textarea-expected.txt

Demonstrates that when typing in a textarea, not all of the InlineTextBoxes need to be updated for every character pressed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS firstInlineTextBoxBefore.isEqual(firstInlineTextBoxAfter) is false
FAIL lastInlineTextBoxBefore.isEqual(lastInlineTextBoxAfter) should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
